+++ components/modals/planificator/shuttle/mixins/passengers-handler.js
@@ -1,12 +1,3 @@
-function resetState() {
-  return {
-    departure: null,
-    arrival: null,
-    email: '',
-    phone: '',
-  };
-}
-
 export default () => ({
   data() {
     return {
@@ -19,33 +10,36 @@ export default () => ({
   methods: {
     addPassenger() {
       if (this.email && this.departure && this.arrival) {
-        let skip = false;
-        this.stops = this.stops.map((stop, index) => {
-          if (this.capacity - stop.passengers.length <= 0) {
-            skip = true;
+        const departure = this.stops.find((stop) => stop.id === this.departure.id);
+        const arrival = this.stops.find((stop) => stop.id === this.arrival.id);
+        if (departure && arrival) {
+          const passenger = {
+            email: this.email,
+            departure: {
+              id: departure.id,
+              label: departure.label,
+            },
+            arrival: {
+              id: arrival.id,
+              label: arrival.label,
+            },
+          };
+
+          if (this.phone) {
+            passenger.phone = this.phone;
           }
-          if (!skip && index >= this.departure.index && index < this.arrival.index) {
-            const passenger = { email: this.email };
-            if (this.phone) {
-              passenger.phone = this.phone;
-            }
-            stop.passengers.push(passenger);
-          }
-          return stop;
-        });
-        resetState();
-        this.$emit('update-stops', this.stops);
+
+          this.passengers.push(passenger);
+        }
+        this.resetState();
+        this.$emit('update-passengers', this.passengers);
       } else {
         this.$toast.error('Veuillez completer l\'email, l\'arret de depart et l\'arret d\'arrivee');
       }
     },
     deletePassenger(passenger) {
-      this.stops = this.stops.map(({ id, label, passengers }) => ({
-        id,
-        label,
-        passengers: passengers.filter((p) => p.email !== passenger.email),
-      }));
-      this.$emit('update-stops', this.stops);
+      this.passengers.splice(this.passengers.findIndex(({ email }) => email === passenger.email), 1);
+      this.$emit('update-passengers', this.passengers);
     },
   },
 });