+++ components/modals/planificator/shuttle/mixins/stops-handler.js
@@ -1,10 +1,7 @@
 export default () => ({
   methods: {
-    updateStops(stops) {
-      this.shuttle.stops = stops;
-    },
-    validatePattern() {
-      this.shuttle.stops = this.shuttle.pattern.stops.map(({ id, label }) => ({
+    validateShuttleFactory() {
+      this.shuttle.stops = this.shuttle.shuttleFactory.stops.map(({ id, label }) => ({
         id,
         label,
         passengers: [],