+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Added crud filter
 * Added root dashboard for multi campus
 * Added capacity to car model crud
-* Added crud for patterns
+* Added crud for shuttle factory
 * Added shuttle licence information to drivers in planning
 * Added shuttle creation and update to planificator
 ### Fix